% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc2.R
\name{AUC_ic2}
\alias{AUC_ic2}
\title{Compute AUC for Interval-Censored Data}
\usage{
AUC_ic2(s, th_v, data3.id, survp, survt, ee)
}
\arguments{
\item{s}{description}

\item{th_v}{description}

\item{data3.id}{A dataset for analysis}

\item{survp}{description}

\item{survt}{description}

\item{ee}{description}
}
\value{
A list containing:
\item{auc1}{}
}
\description{
This function ---
}
