% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{brier}
\alias{brier}
\title{Compute Brier Score for Survival Analysis}
\usage{
brier(data.id, survt, survp, ee, s, th_v)
}
\arguments{
\item{data.id}{A dataset for analysis}

\item{survt}{description}

\item{survp}{description}

\item{ee}{description}

\item{s}{description}

\item{th_v}{description}
}
\value{
A list containing:
\item{bs}{}
}
\description{
This function calculates
}
