% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{data2}
\alias{data2}
\title{Example dataset: data2}
\format{
A data frame with X rows and Y variables:
\describe{
\item{BVRT}{Brief Visuospatial Memory Test score}
\item{CEP}{Cognitive Examination Performance score}
\item{CESD}{Center for Epidemiologic Studies Depression scale score}
\item{HIER}{Hierarchical score}
\item{ID}{Unique subject identifier}
\item{IST}{Intelligence Structure Test score}
\item{MMSE}{Mini-Mental State Examination score}
\item{age}{Current age of the subject}
\item{age_init}{Age at initial assessment}
\item{agedem}{Age at dementia onset (if applicable)}
\item{dem}{Dementia diagnosis (1 = Yes, 0 = No)}
\item{male}{Gender indicator (1 = Male, 0 = Female)}
}
}
\source{
Source of data
}
\usage{
data2
}
\description{
Description of data2.
}
\keyword{datasets}
