% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2id.R
\docType{data}
\name{data2id}
\alias{data2id}
\title{Example dataset: data2id}
\format{
A vector with length X.
}
\source{
Source of data
}
\usage{
data2id
}
\description{
Description of data2id.
}
\keyword{datasets}
